refactor: move /server/relay/ -> /internal/relay/
